+++ scripts/helpers/modulesettings.mjs
@@ -4,6 +4,8 @@ import { MessageParserDnD5e } from '../systemparsers/dnd5e.mjs';
 import { MessageParserPF1 } from '../systemparsers/pf1.mjs';
 import { MessageParserProjectFU } from '../systemparsers/projectfu.mjs';
 import { MessageParserCosmereRPG } from '../systemparsers/cosmere-rpg.mjs';
+import { MessageParserBurningWheel } from '../systemparsers/burningwheel.mjs';
+import { MessageParserCortexPrime } from '../systemparsers/cortexprime.mjs';
 import { ThreadedChatConfig } from '../../src/forms/threadedchatconfig.mjs'
 import { AutoPingConfig } from '../../src/forms/autopingconfig.mjs';
 import { updateServerStatus } from './monitor/serverstatus.mjs';
@@ -381,6 +383,12 @@ export function getSystemParser() {
             break;
         case "cosmere-rpg":
             return new MessageParserCosmereRPG();
+            break;
+        case "burningwheel":
+            return new MessageParserBurningWheel();
+            break;
+        case "cortexprime":
+            return new MessageParserCortexPrime();
             break;
         default:
             systemCompatible = false;